background-worker: Move cleanup code for BackendNotificationPusher in its module
