+++ services/background-worker/src/Wire/BackendNotificationPusher.hs
@@ -22,6 +22,7 @@ import Network.RabbitMqAdmin qualified as RabbitMqAdmin
 import Network.Wai.Utilities.Error
 import Prometheus
 import Servant.Client qualified as Servant
+import System.Logger qualified as Logger
 import System.Logger.Class qualified as Log
 import UnliftIO
 import Wire.API.Federation.API
@@ -314,7 +315,7 @@ getRemoteDomains adminClient = do
           . Log.field "queue" ("backend-notifications." <> d)
           . Log.field "error" e
 
-startWorker :: AmqpEndpoint -> AppT IO (IORef (Maybe Q.Channel), IORef (Map Domain (Q.ConsumerTag, MVar ())))
+startWorker :: AmqpEndpoint -> AppT IO CleanupAction
 startWorker rabbitmqOpts = do
   env <- ask
   -- These are used in the POSIX signal handlers, so we need to make
@@ -352,4 +353,28 @@ startWorker rabbitmqOpts = do
                     clearRefs
                     runAppT env $ markAsNotWorking BackendNotificationPusher
                 }
-  pure (chanRef, consumersRef)
+  pure $ cleanupWorker env.logger chanRef consumersRef
+
+cleanupWorker :: Logger.Logger -> IORef (Maybe Q.Channel) -> IORef (Map Domain (Q.ConsumerTag, MVar ())) -> CleanupAction
+cleanupWorker l notifChanRef notifConsumersRef = do
+  Logger.info l $ Log.msg (Log.val "Cancelling the notification pusher thread")
+  readIORef notifChanRef >>= traverse_ \chan -> do
+    Logger.info l $ Log.msg (Log.val "Got channel")
+    readIORef notifConsumersRef >>= \m -> for_ (Map.assocs m) \(domain, (consumer, runningFlag)) -> do
+      Logger.info l $ Log.msg (Log.val "Cancelling consumer") . Log.field "Domain" domain._domainText
+      -- Remove the consumer from the channel so it isn't called again
+      Q.cancelConsumer chan consumer
+      -- Take from the mvar. This will only unblock when the consumer callback isn't running.
+      -- This allows us to wait until the currently running tasks are completed, and new ones
+      -- won't be scheduled because we've already removed the callback from the channel.
+      -- If, for some reason, a consumer is invoked after us cancelling it, taking this MVar
+      -- will block that thread from trying to push out the notification. At this point, we're
+      -- relying on Rabbit to requeue the message for us as we won't be able to ACK or NACK it.
+      -- This helps prevent message redelivery to endpoint services during the brief window between
+      -- receiving a message from rabbit, and the signal handler shutting down the AMQP connection
+      -- before notification delivery has finalised.
+      Logger.info l $ Log.msg $ Log.val "Taking MVar. Waiting for current operation to finish"
+      takeMVar runningFlag
+    -- Close the channel. `extended` will then close the connection, flushing messages to the server.
+    Logger.info l $ Log.msg $ Log.val "Closing RabbitMQ channel"
+    Q.closeChannel chan
